Add +dfsg suffix to the upstream version.
We remove files from the original tarball with d/copyright's
Files-Excluded for DFSG reasons (upstream ships generated files,
which we can't prove to be built from the shipped source
otherwise).  So indicate that this happened in the version number
(if no files were to be removed, the suffix wouldn't be added by
this change).

Wrap the opts.