Change watch URL to consider all Wine releases.
The Wine stable release candidates are in the .0 directories on
the Wine download server, but currently they are packaged in
wine-development which tracks the .x directories.  This commit
changes the regex to any character, so stable releases and their
release candidates are considered versions for wine-development.

Also allow for multi-digit Wine releases and escape the dot in
the version regex.