Move the wineserver script from wine to libwine.
Closes: #932201

This ensures that Wine always finds its wineserver in its bindir.
Otherwise if wine was not installed it might fail to find it, or
fall back to another wineserver in e.g. PATH.

wine{32,64} would be better suited, but dpkg only accepts multiple
packages as owner of a file for "Multi-Arch: same" packages.